Crud y Login Funcional

+++ src/main/java/com/project/company/management/repositories/IRepositoryTransaction.java
@@ -13,7 +13,7 @@ public interface IRepositoryTransaction extends CrudRepository<Transaction, Long
 
     //Total Monto de movimientos de una empresa especifica
     @Query(value = "select sum(amount) from Transaction where enterprise_id= ?1",nativeQuery = true)
-    public abstract float totalAmount(long id);
+    public abstract Long totalAmount(long id);
 
     //Guardar movimiento de una empresa especifica
     @Query(value = "select * from Transaction where enterprise_id = ?1 and id = ?2", nativeQuery = true)
